fix: moving multisig utils
