+++ src/execution/utils.ts
@@ -172,54 +172,3 @@ export function checkStepSlippageThreshold(
     actualSlippage.lte(setSlippage)
   )
 }
-
-export const updateSafeRouteProcess = async (
-  internalTxHash: string,
-  step: LifiStep,
-  statusManager: StatusManager,
-  process: Process,
-  fromChain: ExtendedChain
-) => {
-  const config = ConfigService.getInstance().getConfig()
-
-  if (!config.multisigConfig?.getMultisigTransactionDetails) {
-    throw new Error(
-      '"getMultisigTransactionDetails()" is missing in Multisig config.'
-    )
-  }
-
-  const safeStatusResponse: MultisigTxDetails =
-    await config.multisigConfig?.getMultisigTransactionDetails(
-      internalTxHash,
-      fromChain.id
-    )
-
-  if (safeStatusResponse.status === 'DONE') {
-    console.log('Updating transaction hash --> ', {
-      txHash: safeStatusResponse.txHash,
-    })
-
-    process = statusManager.updateProcess(step, process.type, 'PENDING', {
-      txHash: safeStatusResponse.txHash,
-      multisigTxHash: undefined,
-      txLink:
-        fromChain.metamask.blockExplorerUrls[0] +
-        'tx/' +
-        safeStatusResponse.txHash,
-    })
-  }
-
-  if (safeStatusResponse.status === 'FAILED') {
-    throw new TransactionError(
-      LifiErrorCode.TransactionFailed,
-      'Multisig transaction failed.'
-    )
-  }
-
-  if (safeStatusResponse.status === 'CANCELLED') {
-    throw new TransactionError(
-      LifiErrorCode.TransactionRejected,
-      'Transaction was rejected by users.'
-    )
-  }
-}
